Added a new Section to bit-manipulation.md
There is a shortcut method to count the total number of set bits in all the numbers from 1 upto n.
This method is added to the as a small Trick  to the Useful Tricks section of the file after the "Brian Kernighan's algorithm" section, which counts the no. of set bits in a given number.